dts: rp1: Don't use DMA with UARTs
DMA has been enabled on RP1's UART0, but with mixed success. Transmits
seem to work, but the DMA interface is not well suited to receiving
arbitrary amounts of data. In particular, the PL011 driver is slow to
pass on the received data, batching it into large blocks.

On balance, it's better to just disable the DMA support. As with the
other UARTs, the required runes are left in the DTS as comments.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>